refactor: module structures
Registers are not only for ahc, but also for ports.
